Finish main functions

+++ GBCLV3/Services/Launcher/AssetService.cs
@@ -45,7 +45,7 @@ namespace GBCLV3.Services.Launcher
 
         public bool LoadAllObjects(AssetsInfo info)
         {
-            string jsonPath = $"{_gamePathService.AssetDir}/indexes/{info.ID}.json";
+            string jsonPath = $"{_gamePathService.AssetsDir}/indexes/{info.ID}.json";
             if (!File.Exists(jsonPath)) return false;
 
             if (info.Objects != null) return true;
@@ -67,7 +67,7 @@ namespace GBCLV3.Services.Launcher
                 .AsParallel()
                 .Where(obj =>
                 {
-                    string path = $"{_gamePathService.AssetDir}/objects/{obj.Path}";
+                    string path = $"{_gamePathService.AssetsDir}/objects/{obj.Path}";
                     return !File.Exists(path) || obj.Hash != Utils.CryptUtil.GetFileSHA1(path);
                 })
                 .ToList()
@@ -84,17 +84,14 @@ namespace GBCLV3.Services.Launcher
                 .AsParallel()
                 .ForAll(pair =>
                 {
-                    var objectPath = $"{_gamePathService.AssetDir}/objects/{pair.Value.Path}";
-                    var virtualPath = $"{_gamePathService.AssetDir}/virtual/legacy/{pair.Key}";
+                    var objectPath = $"{_gamePathService.AssetsDir}/objects/{pair.Value.Path}";
+                    var virtualPath = $"{_gamePathService.AssetsDir}/virtual/legacy/{pair.Key}";
                     var virtualDir = Path.GetDirectoryName(virtualPath);
 
-                    if (File.Exists(virtualPath)) return;
+                    if (!File.Exists(objectPath) || File.Exists(virtualPath)) return;
 
-                    if (!Directory.Exists(virtualDir))
-                    {
-                        Directory.CreateDirectory(virtualDir);
-                    }
-
+                    // Make sure directory exists
+                    Directory.CreateDirectory(virtualDir);
                     File.Copy(objectPath, virtualPath);
                 })
             );
@@ -105,13 +102,10 @@ namespace GBCLV3.Services.Launcher
             try
             {
                 var json = await _client.GetStringAsync(_urlService.Base.Json + info.IndexUrl);
-                var indexDir = $"{_gamePathService.AssetDir}/indexes";
+                var indexDir = $"{_gamePathService.AssetsDir}/indexes";
 
-                if (!Directory.Exists(indexDir))
-                {
-                    Directory.CreateDirectory(indexDir);
-                }
-
+                //Make sure directory exists
+                Directory.CreateDirectory(indexDir);
                 File.WriteAllText($"{indexDir}/{info.ID}.json", json, Encoding.UTF8);
                 return true;
             }
@@ -127,19 +121,18 @@ namespace GBCLV3.Services.Launcher
             }
         }
 
-        public (DownloadType, IEnumerable<DownloadItem>) GetDownloadInfo(IEnumerable<AssetObject> assetObjects)
+        public IEnumerable<DownloadItem> GetDownloads(IEnumerable<AssetObject> assetObjects)
         {
-            var items = assetObjects.Select(obj => new DownloadItem
+            return assetObjects.Select(obj => 
+            new DownloadItem
             {
                 Name = obj.Hash,
-                Path = $"{_gamePathService.AssetDir}/objects/{obj.Path}",
+                Path = $"{_gamePathService.AssetsDir}/objects/{obj.Path}",
                 Url = _urlService.Base.Asset + obj.Path,
                 Size = obj.Size,
                 IsCompleted = false,
                 DownloadedBytes = 0,
             }).ToList();
-
-            return (DownloadType.Assets, items);
         }
 
         #endregion
